перенес вспомогательные функции в utils, откорректировал модули, тесты, фикстуры
